+++ packages/spacecat-shared-data-access/test/it/scrape-job/scrape-job.test.js
@@ -16,6 +16,7 @@ import chaiAsPromised from 'chai-as-promised';
 import ScrapeJobModel from '../../../src/models/scrape-job/scrape-job.model.js';
 import { getDataAccess } from '../util/db.js';
 import { seedDatabase } from '../util/seed.js';
+import { IT_HOOK_TIMEOUT } from '../util/util.js';
 
 use(chaiAsPromised);
 
@@ -43,7 +44,7 @@ describe('ScrapeJob IT', async () => {
   let newJobData;
 
   before(async function () {
-    this.timeout(10000);
+    this.timeout(IT_HOOK_TIMEOUT);
     sampleData = await seedDatabase();
 
     const dataAccess = getDataAccess();