+++ packages/spacecat-shared-data-access/test/it/site-ims-org-access/site-ims-org-access.test.js
@@ -18,6 +18,7 @@ import DataAccessError from '../../../src/errors/data-access.error.js';
 import SiteImsOrgAccessCollection from '../../../src/models/site-ims-org-access/site-ims-org-access.collection.js';
 import { getDataAccess } from '../util/db.js';
 import { seedDatabase } from '../util/seed.js';
+import { IT_HOOK_TIMEOUT } from '../util/util.js';
 
 use(chaiAsPromised);
 
@@ -26,7 +27,7 @@ describe('SiteImsOrgAccess IT', async () => {
   let SiteImsOrgAccess;
 
   before(async function () {
-    this.timeout(10000);
+    this.timeout(IT_HOOK_TIMEOUT);
     sampleData = await seedDatabase();
 
     const dataAccess = getDataAccess();